+++ libs/web/src/grids/groups-grid/group-select-input/group-select-input.component.ts
@@ -1,17 +1,19 @@
+import 'rxjs/add/operator/takeUntil';
+
 import { Component, ComponentFactoryResolver, EventEmitter, Input, Output, ViewChild } from '@angular/core';
 import { TranslateService } from '@ngx-translate/core';
-import { Group } from '@rucken/core';
-import { User } from '@rucken/core';
-import { GroupsService } from '@rucken/core';
 import { AppService } from '@rucken/core';
+import { GroupsService } from '@rucken/core';
+import { User } from '@rucken/core';
+import { Group } from '@rucken/core';
 import { AccountService } from '@rucken/core';
 import { TooltipDirective } from 'ngx-bootstrap/tooltip';
 
 import {
-    BaseResourceSelectInputComponent,
+  BaseResourceSelectInputComponent,
 } from './../../../base/base-resources-grid/base-resource-select-input/base-resource-select-input.component';
 import {
-    BaseResourceSelectInputConfig,
+  BaseResourceSelectInputConfig,
 } from './../../../base/base-resources-grid/base-resource-select-input/base-resource-select-input.config';
 import { SelectInputComponent } from './../../../controls/select-input/select-input.component';
 import { GroupsListModalComponent } from './../groups-list-modal/groups-list-modal.component';
@@ -49,13 +51,7 @@ export class GroupSelectInputComponent extends BaseResourceSelectInputComponent 
     public config: BaseResourceSelectInputConfig
   ) {
     super(translateService, config);
-  }
-  afterCreate() {
-    super.afterCreate();
     this.cachedResourcesService = this.groupsService.createCache();
-  }
-  get account(): any | User {
-    return this.accountService.account;
   }
   onLookup() {
     const itemModal: GroupsListModalComponent =
@@ -78,7 +74,7 @@ export class GroupSelectInputComponent extends BaseResourceSelectInputComponent 
     itemModal.onClose.subscribe(() => this.focus());
     itemModal.item = this.value;
     itemModal.modal.show();
-    this.cachedResourcesService.changeStatusItem$.subscribe(status =>
+    this.cachedResourcesService.changeStatusItem$.takeUntil(this.destroyed$).subscribe(status =>
       itemModal.okInProcessFromStatus(status)
     );
   }