+++ libs/web/src/grids/groups-grid/groups-grid.component.ts
@@ -1,3 +1,5 @@
+import 'rxjs/add/operator/takeUntil';
+
 import { Component, ComponentFactoryResolver, ElementRef, EventEmitter, Output, ViewChild } from '@angular/core';
 import { TranslateService } from '@ngx-translate/core';
 import { User } from '@rucken/core';
@@ -40,8 +42,8 @@ export class GroupsGridComponent extends BaseResourcesGridComponent {
     super();
     this.cachedResourcesService = this.groupsService.createCache();
   }
-  get account(): any | User {
-    return this.accountService.account;
+  init() {
+    super.init();
   }
   get readonly() {
     return this.hardReadonly || !this.account || !this.account.checkPermissions(['add_group', 'change_group', 'delete_group']);
@@ -62,7 +64,7 @@ export class GroupsGridComponent extends BaseResourcesGridComponent {
     itemModal.item = new Group();
     itemModal.modal.show();
     this.selectedItems = [itemModal.item];
-    this.cachedResourcesService.changeStatusItem$.subscribe(status =>
+    this.cachedResourcesService.changeStatusItem$.takeUntil(this.destroyed$).subscribe(status =>
       itemModal.okInProcessFromStatus(status)
     );
   }
@@ -85,7 +87,7 @@ export class GroupsGridComponent extends BaseResourcesGridComponent {
     itemModal.item = new Group(item);
     itemModal.modal.show();
     this.selectedItems = [itemModal.item];
-    this.cachedResourcesService.changeStatusItem$.subscribe(status =>
+    this.cachedResourcesService.changeStatusItem$.takeUntil(this.destroyed$).subscribe(status =>
       itemModal.okInProcessFromStatus(status)
     );
   }
@@ -103,7 +105,7 @@ export class GroupsGridComponent extends BaseResourcesGridComponent {
     confirm.onClose.subscribe(() => this.focus());
     this.selectedItems = [item];
     confirm.modal.show();
-    this.cachedResourcesService.changeStatusItem$.subscribe(status =>
+    this.cachedResourcesService.changeStatusItem$.takeUntil(this.destroyed$).subscribe(status =>
       confirm.okInProcessFromStatus(status)
     );
   }